Some fun with future

+++ task5 - Future Java/src/main/java/com/codecool/spingboot_tasks/future_java/controller/LongCalculationsController.java
@@ -4,6 +4,8 @@ import com.codecool.spingboot_tasks.future_java.service.LongCalculationsService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigInteger;
+
 @RestController
 public class LongCalculationsController {
 
@@ -14,7 +16,7 @@ public class LongCalculationsController {
     }
 
     @GetMapping("/calc")
-    public int calc(){
+    public BigInteger calc(){
         return service.calc();
     }
 }
